BASIC ORDER CREATION and RETREIVING DONE

+++ src/main/java/com/rgiftings/Backend/DTO/Product/Create/ProductAttributeValueRequest.java
@@ -1,7 +1,9 @@
 package com.rgiftings.Backend.DTO.Product.Create;
 
+import java.math.BigDecimal;
+
 public record ProductAttributeValueRequest(
         Long attributeValueId,
-        Double extraPrice
+        BigDecimal extraPrice
 ) {
 }
