+++ src/main/java/com/rgiftings/Backend/DTO/Product/Retrieve/ProductAttributeValueResponse.java
@@ -1,9 +1,11 @@
 package com.rgiftings.Backend.DTO.Product.Retrieve;
 
+import java.math.BigDecimal;
+
 public record ProductAttributeValueResponse(
         Long id,
         Long valueId,
         String valueName,
-        Double extraPrice
+        BigDecimal extraPrice
 ) {
 }
