+++ src/main/java/com/rgiftings/Backend/DTO/Product/Update/UpdateProductAttributeValueRequest.java
@@ -1,8 +1,10 @@
 package com.rgiftings.Backend.DTO.Product.Update;
 
+import java.math.BigDecimal;
+
 public record UpdateProductAttributeValueRequest(
         Long id,
         Long AttributeValueId,
-        Double extraPrice
+        BigDecimal extraPrice
 ) {
 }
